pyproject.toml
[tool.poetry]
name = "weather-project-v1-1"
version = "0.1.0"
description = ""
authors = ["myuR <myurpvn@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.3"
requests = "^2.31.0"
google-cloud-bigquery = "^3.13.0"
pyarrow = "^14.0.1"
numpy = "^1.26.2"
structlog = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
